app/backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Sample crud with fast api"
authors = ["alexandreFialho <alexandrefialhobr@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "0.63.0"
uvicorn = "0.13.4"
requests = "2.25.1"
pydantic = "1.8.1"
python-jose = {extras = ["cryptography"], version = "^3.1.0"}
python-multipart = "0.0.5"
passlib = {extras = ["bcrypt"], version = "1.7.4"}
pycep-correios = "5.0.0"
SQLAlchemy = "1.3.16"
SQLAlchemy-Utils = "0.36.8"
alembic = "1.5.6"
psycopg2-binary = "2.8.6"
statsd = "^3.3.0"

[tool.poetry.dev-dependencies]
pytest = "6.2.2"
ipython = "7.21.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
